requirements.txt
langchain==0.1.4 # https://pypi.org/project/langchain/
langchainhub==0.1.14
llama-index==0.9.39
ollama==0.1.4 #https://pypi.org/project/ollama/
langchain-experimental==0.0.49 #https://pypi.org/project/langchain-experimental/
langchain-community

qdrant-client
sentence-transformers
EbookLib
html2text

# Observability
arize-phoenix

# Agent tools
requests
duckduckgo_search


# PDF tools
pypdf

# Quality of life tooling
python-dotenv
rich

# For multi-modal models
pillow